+++ docs/conf.py
@@ -1,4 +1,3 @@
-# -*- coding: utf-8 -*-
 #
 # Flask Dance documentation build configuration file, created by
 # sphinx-quickstart on Sat Sep 27 09:47:52 2014.
@@ -208,8 +207,8 @@ latex_documents = [
     (
         "index",
         "FlaskDance.tex",
-        u"Flask Dance Documentation",
-        u"David Baumgold",
+        "Flask Dance Documentation",
+        "David Baumgold",
         "manual",
     )
 ]
@@ -240,7 +239,7 @@ latex_logo = "_static/flask-dance.png"
 # One entry per manual page. List of tuples
 # (source start file, name, description, authors, manual section).
 man_pages = [
-    ("index", "flaskdance", u"Flask Dance Documentation", [u"David Baumgold"], 1)
+    ("index", "flaskdance", "Flask Dance Documentation", ["David Baumgold"], 1)
 ]
 
 # If true, show URL addresses after external links.
@@ -256,8 +255,8 @@ texinfo_documents = [
     (
         "index",
         "FlaskDance",
-        u"Flask Dance Documentation",
-        u"David Baumgold",
+        "Flask Dance Documentation",
+        "David Baumgold",
         "FlaskDance",
         "One line description of project.",
         "Miscellaneous",