+++ flask_dance/__init__.py
@@ -1,6 +1,3 @@
-# coding=utf-8
-from __future__ import unicode_literals
-
 from .consumer import OAuth1ConsumerBlueprint, OAuth2ConsumerBlueprint
 
 __version__ = "3.3.0"