+++ flask_dance/contrib/authentiq.py
@@ -1,5 +1,3 @@
-from __future__ import unicode_literals
-
 from flask_dance.consumer import OAuth2ConsumerBlueprint
 from functools import partial
 from flask.globals import LocalProxy, _lookup_app_object
@@ -63,9 +61,9 @@ def make_authentiq_blueprint(
         client_id=client_id,
         client_secret=client_secret,
         scope=scope,
-        base_url="https://{hostname}/".format(hostname=hostname),
-        authorization_url="https://{hostname}/authorize".format(hostname=hostname),
-        token_url="https://{hostname}/token".format(hostname=hostname),
+        base_url=f"https://{hostname}/",
+        authorization_url=f"https://{hostname}/authorize",
+        token_url=f"https://{hostname}/token",
         redirect_url=redirect_url,
         redirect_to=redirect_to,
         login_url=login_url,